refactor: Iteration function reworked
Previously the middle iteration was always calculated for 1000 steps,
which is just a  waste of time. It is now reworked in such a way that
the Iteration Function stops if the difference between two steps smaller
than 1e-8, or if it reached 1000 iteration. In latter case it raises
error.

Modified featers:
 - Iteration function
 - Added an IterationError class
 - Added difference calculation