feat/refactor: Newton iteration added
In the previus versions it used a 6 step iteration for the calculations,
which wasn't accurate. In this commit I implemented multi-variable
newton-iteration.
The first two steps gives a good approximation for the starting x,y and
z values, and the implemented Jacobian and function vectors do the
iteration.

It needs refactoring:
 - clean up old iterations
 - Extract function for newtonian
 - Create class for iteration
 - Refactor Constants class