refactor: Ionization, NewtonianIterator classes
Achieving the main goal, that is to use OOP interfaces between tcdata
and calcion. It was necesseary to refactor the previous functions into
function methods.
Since logically somewhat distinct, I use two new classes: one for
the calculation, and one for the newton iterations.
It will still need docs and the actual interface for the
tcdata.DataPoint object.